Перенес методы в отдельные файлы и сгруппировал их в папки
