--- GBShop/App/BusinessLogic/Network/Auth.swift
@@ -1,152 +1,0 @@
-//
-//  Auth.swift
-//  GBShop
-//
-//  Created by Alexander Myskin on 13.02.2021.
-//
-
-import Foundation
-import Alamofire
-
-class Auth: AbstractRequestFactory {
-    let errorParser: AbstractErrorParser
-    let sessionManager: Session
-    let queue: DispatchQueue
-    let baseUrl = URL(string: "https://raw.githubusercontent.com/GeekBrainsTutorial/online-store-api/master/responses/")!
-    
-    init(
-        errorParser: AbstractErrorParser,
-        sessionManager: Session,
-        queue: DispatchQueue = DispatchQueue.global(qos: .utility)) {
-        self.errorParser = errorParser
-        self.sessionManager = sessionManager
-        self.queue = queue
-    }
-}
-
-extension Auth: AuthRequestFactory {
-    func login(userName: String, password: String, completionHandler: @escaping (AFDataResponse<LoginResult>) -> Void) {
-        let requestModel = Login(baseUrl: baseUrl, login: userName, password: password)
-        self.request(request: requestModel, completionHandler: completionHandler)
-    }
-    func logout(userId: Int, completionHandler: @escaping (AFDataResponse<LogoutResult>) -> Void) {
-        let requestModel = Logout(baseUrl: baseUrl, idUser: userId)
-        self.request(request: requestModel, completionHandler: completionHandler)
-    }
-    func registration(userId: Int,
-                      userName: String,
-                      password: String,
-                      email: String,
-                      gender: String,
-                      creditCard: String,
-                      bio: String,
-                      completionHandler: @escaping (AFDataResponse<RegistrationResult>) -> Void) {
-        let requestModel = Registeration(baseUrl: baseUrl,
-                                         idUser: userId,
-                                         login: userName,
-                                         password: password,
-                                         email: email,
-                                         gender: gender,
-                                         creditCard: creditCard,
-                                         bio: bio
-                                         )
-        self.request(request: requestModel, completionHandler: completionHandler)
-    }
-    func changeUserData(userId: Int,
-                      userName: String,
-                      password: String,
-                      email: String,
-                      gender: String,
-                      creditCard: String,
-                      bio: String,
-                      completionHandler: @escaping (AFDataResponse<ChangeUserResult>) -> Void) {
-        let requestModel = Registeration(baseUrl: baseUrl,
-                                         idUser: userId,
-                                         login: userName,
-                                         password: password,
-                                         email: email,
-                                         gender: gender,
-                                         creditCard: creditCard,
-                                         bio: bio
-                                         )
-        self.request(request: requestModel, completionHandler: completionHandler)
-    }
-}
-
-extension Auth {
-    struct Login: RequestRouter {
-        let baseUrl: URL
-        let method: HTTPMethod = .get
-        let path: String = "login.json"
-        
-        let login: String
-        let password: String
-        var parameters: Parameters? {
-            return [
-                "username": login,
-                "password": password
-            ]
-        }
-    }
-    struct Logout: RequestRouter {
-        let baseUrl: URL
-        let method: HTTPMethod = .get
-        let path: String = "logout.json"
-        
-        let idUser: Int
-        var parameters: Parameters? {
-            return [
-                "id_user": idUser
-            ]
-        }
-    }
-    struct Registeration: RequestRouter {
-        let baseUrl: URL
-        let method: HTTPMethod = .get
-        let path: String = "registerUser.json"
-        
-        let idUser: Int
-        let login: String
-        let password: String
-        let email: String
-        let gender: String
-        let creditCard: String
-        let bio: String
-        var parameters: Parameters? {
-            return [
-                "id_user": idUser,
-                "username": login,
-                "password": password,
-                "email": email,
-                "gender": gender,
-                "credit_card": creditCard,
-                "bio": bio
-            ]
-        }
-    }
-    
-    struct ChangeUserData: RequestRouter {
-        let baseUrl: URL
-        let method: HTTPMethod = .get
-        let path: String = "changeUserData.json"
-        
-        let idUser: Int
-        let login: String
-        let password: String
-        let email: String
-        let gender: String
-        let creditCard: String
-        let bio: String
-        var parameters: Parameters? {
-            return [
-                "id_user": idUser,
-                "username": login,
-                "password": password,
-                "email": email,
-                "gender": gender,
-                "credit_card": creditCard,
-                "bio": bio
-            ]
-        }
-    }
-}
