+++ GBShop/App/BusinessLogic/Network/Auth/Auth.swift
@@ -1,0 +1,50 @@
+//
+//  Auth.swift
+//  GBShop
+//
+//  Created by Alexander Myskin on 13.02.2021.
+//
+
+import Foundation
+import Alamofire
+
+class Auth: AbstractRequestFactory {
+    let errorParser: AbstractErrorParser
+    let sessionManager: Session
+    let queue: DispatchQueue
+    let baseUrl = URL(string: AppApi.myURL.rawValue)!
+    
+    init(
+        errorParser: AbstractErrorParser,
+        sessionManager: Session,
+        queue: DispatchQueue = DispatchQueue.global(qos: .utility)) {
+        self.errorParser = errorParser
+        self.sessionManager = sessionManager
+        self.queue = queue
+        
+    }
+}
+
+extension Auth: AuthRequestFactory {
+    func login(userName: String, password: String, completionHandler: @escaping (AFDataResponse<LoginResult>) -> Void) {
+        let requestModel = Login(baseUrl: baseUrl, login: userName, password: password)
+        self.request(request: requestModel, completionHandler: completionHandler)
+    }
+}
+
+extension Auth {
+    struct Login: RequestRouter {
+        let baseUrl: URL
+        let method: HTTPMethod = .get
+        let path: String = "login.json"
+        
+        let login: String
+        let password: String
+        var parameters: Parameters? {
+            return [
+                "username": login,
+                "password": password
+            ]
+        }
+    }
+}
