--- GBShop/App/BusinessLogic/Network/AuthRequestFactory.swift
@@ -1,29 +1,0 @@
-//
-//  AuthRequestFactory.swift
-//  GBShop
-//
-//  Created by Alexander Myskin on 13.02.2021.
-//
-
-import Alamofire
-
-protocol AuthRequestFactory {
-    func login(userName: String, password: String, completionHandler: @escaping (AFDataResponse<LoginResult>) -> Void)
-    func logout(userId: Int, completionHandler: @escaping (AFDataResponse<LogoutResult>) -> Void)
-    func registration(userId: Int,
-                      userName: String,
-                      password: String,
-                      email: String,
-                      gender: String,
-                      creditCard: String,
-                      bio: String,
-                      completionHandler: @escaping (AFDataResponse<RegistrationResult>) -> Void)
-    func changeUserData(userId: Int,
-                      userName: String,
-                      password: String,
-                      email: String,
-                      gender: String,
-                      creditCard: String,
-                      bio: String,
-                      completionHandler: @escaping (AFDataResponse<ChangeUserResult>) -> Void)
-}
