+++ GBShop/App/BusinessLogic/Network/BaseURL/BaseURL.swift
@@ -1,0 +1,17 @@
+//
+//  AppEndPoint.swift
+//  GBShop
+//
+//  Created by Alexander Myskin on 16.02.2021.
+//
+
+import Foundation
+
+
+
+enum AppApi: String {
+    case myURL = "https://raw.githubusercontent.com/GeekBrainsTutorial/online-store-api/master/responses/"
+}
+
+
+
